Lint agregado

+++ lib/domain/cubits/groups/group_details/group_details_cubit.dart
@@ -26,7 +26,7 @@ class GroupDetailsCubit extends Cubit<GroupDetailsState> {
         actions.add(transaction);
       }
 
-      actions.sort((a, b) => b.timestamp.compareTo(a.timestamp));
+      actions.sort((a, b) => b.timestamp.compareTo(a.timestamp) as int);
 
       emit(GroupDetailsLoaded(_group, actions));
     });
